add expense page(finished), add category page(still not finished)

+++ lib/logic/providers/budget_provider.dart
@@ -16,7 +16,7 @@ class BudgetProvider extends ChangeNotifier {
     notifyListeners();
   }
 
-  void increaseSpentBudget(int addedMoney) {
+  void increaseSpentBudget(double addedMoney) {
     model.spentBudget += addedMoney;
     notifyListeners();
   }
